+++ cloudsmith_cli/cli/commands/credential_helper/__init__.py
@@ -10,6 +10,7 @@ import click
 
 from ..main import main
 from .docker import docker as docker_cmd
+from .generic import generic as generic_cmd
 from .manage import install_cmd, list_cmd, uninstall_cmd
 from .shell import shell_init
 
@@ -33,6 +34,7 @@ def credential_helper():
 
 
 credential_helper.add_command(docker_cmd, name="docker")
+credential_helper.add_command(generic_cmd, name="generic")
 credential_helper.add_command(shell_init, name="shell-init")
 credential_helper.add_command(install_cmd, name="install")
 credential_helper.add_command(uninstall_cmd, name="uninstall")